cleaning and testing ft_export. Not finished yet
