+++ Makefile
@@ -6,7 +6,7 @@
 #    By: vjan-nie <vjan-nie@student.42madrid.com    +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2025/01/17 12:24:16 by vjan-nie          #+#    #+#              #
-#    Updated: 2025/09/25 13:40:48 by vjan-nie         ###   ########.fr        #
+#    Updated: 2025/09/28 11:58:52 by vjan-nie         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -36,6 +36,7 @@ SRC		=	$(SRC_DIR)/minishell.c\
 			$(SRC_DIR)/environment.c\
 			$(SRC_DIR)/environment_utils.c\
 			$(SRC_DIR)/environment_utils2.c\
+			$(SRC_DIR)/environment_utils3.c\
 			$(SRC_DIR)/lexer.c\
 			$(SRC_DIR)/lexer_tokens.c\
 			$(SRC_DIR)/lexer_tokens2.c\
@@ -58,12 +59,12 @@ SRC		=	$(SRC_DIR)/minishell.c\
 			$(SRC_DIR)/redir.c\
 			$(SRC_DIR)/redir_utils.c\
 			$(SRC_DIR)/here_doc.c\
-			$(SRC_DIR)/built_ins.c\
-			$(SRC_DIR)/built_ins_utils.c\
-			$(SRC_DIR)/exit.c\
-			$(SRC_DIR)/echo.c\
-			$(SRC_DIR)/cd.c\
-			$(SRC_DIR)/export.c\
+			$(SRC_DIR)/builtins.c\
+			$(SRC_DIR)/builtins_utils.c\
+			$(SRC_DIR)/builtins_exit.c\
+			$(SRC_DIR)/builtins_echo.c\
+			$(SRC_DIR)/builtins_cd.c\
+			$(SRC_DIR)/builtins_export.c\
 			$(SRC_DIR)/pipes.c\
 			$(SRC_DIR)/pipes_utils.c
