+++ src/builtins_echo.c
